Move app install pages to app router
