+++ apps/web/app/future/apps/installation/[step]/page.tsx
@@ -1,4 +1,3 @@
-import type { OnboardingPageProps } from "@pages/apps/installation/[[...step]]";
 import Page from "@pages/apps/installation/[[...step]]";
 import { withAppDirSsr } from "app/WithAppDirSsr";
 import { _generateMetadata } from "app/_utils";
@@ -7,6 +6,8 @@ import { WithLayout } from "app/layoutHOC";
 import { APP_NAME } from "@calcom/lib/constants";
 
 import { getServerSideProps } from "@lib/apps/installation/getServerSideProps";
+
+import type { OnboardingPageProps } from "~/apps/installation/lib/types";
 
 export const generateMetadata = async () => {
   return await _generateMetadata(
