+++ apps/web/components/apps/installation/EventTypeAppSettingsWrapper.tsx
@@ -1,4 +1,3 @@
-import type { TEventType } from "@pages/apps/installation/[[...step]]";
 import { useEffect, type FC } from "react";
 
 import { EventTypeAppSettings } from "@calcom/app-store/_components/EventTypeAppSettingsInterface";
@@ -7,6 +6,8 @@ import type { EventTypeAppsList } from "@calcom/app-store/utils";
 import useAppsData from "@lib/hooks/useAppsData";
 
 import type { ConfigureStepCardProps } from "@components/apps/installation/ConfigureStepCard";
+
+import type { TEventType } from "~/apps/installation/lib/types";
 
 type EventTypeAppSettingsWrapperProps = Pick<
   ConfigureStepCardProps,
