+++ apps/web/components/apps/installation/EventTypesStepCard.tsx
@@ -1,4 +1,3 @@
-import type { TEventType, TEventTypesForm } from "@pages/apps/installation/[[...step]]";
 import type { Dispatch, SetStateAction } from "react";
 import type { FC } from "react";
 import React from "react";
@@ -7,6 +6,8 @@ import { useFieldArray, useFormContext } from "react-hook-form";
 import { useLocale } from "@calcom/lib/hooks/useLocale";
 import { EventTypeMetaDataSchema } from "@calcom/prisma/zod-utils";
 import { ScrollableArea, Badge, Button } from "@calcom/ui";
+
+import type { TEventType, TEventTypesForm } from "~/apps/installation/lib/types";
 
 type EventTypesCardProps = {
   userName: string;
