+++ apps/web/modules/apps/installation/lib/steps.ts
@@ -1,0 +1,7 @@
+import { AppOnboardingSteps } from "@calcom/lib/apps/appOnboardingSteps";
+
+export const STEPS = [
+  AppOnboardingSteps.ACCOUNTS_STEP,
+  AppOnboardingSteps.EVENT_TYPES_STEP,
+  AppOnboardingSteps.CONFIGURE_STEP,
+] as const;
